requirements.txt
pandas
spotipy
requests
psycopg2-binary
python-dotenv